+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog][docs-changelog], and the version adher
 
 
 ## Unreleased
+### Fixed
+- Invalid InstanceStatus enum value.
 
 ## [0.2.0][changes-0.2.0] - 2025-12-03
 ### Added