+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog][docs-changelog], and the version adher
 
 
 ## Unreleased
+### Added
+- Timeout to cluster / cluster-set status fetching.
 ### Fixed
 - Invalid InstanceStatus enum value.