+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog][docs-changelog], and the version adher
 
 
 ## Unreleased
+
+## [0.3.0][changes-0.3.0] - 2025-12-05
 ### Added
 - Process search method to InstanceClient class
 - Reload TLS method to InstanceClient class.
@@ -35,5 +37,6 @@ The format is based on [Keep a Changelog][docs-changelog], and the version adher
 
 [changes-0.1.0]: https://github.com/canonical/mysql-shell-client/releases/tag/0.1.0
 [changes-0.2.0]: https://github.com/canonical/mysql-shell-client/compare/0.1.0...0.2.0
+[changes-0.3.0]: https://github.com/canonical/mysql-shell-client/compare/0.2.0...0.3.0
 [docs-changelog]: https://keepachangelog.com/en/1.0.0/
 [docs-semver]: https://semver.org/spec/v2.0.0.html